SystemTime can be int... no need for it to be long.

+++ src/globals.h
@@ -79,7 +79,7 @@ EXTERN  INIT(   int     SortByDate, 0);
 EXTERN	INIT(	int	SortByPrio, 0);
 EXTERN  INIT(   int     UntimedBeforeTimed, 0);
 EXTERN	INIT(	int	DefaultPrio, NO_PRIORITY);
-EXTERN  INIT(   long    SysTime, -1L);
+EXTERN  INIT(   int     SysTime, -1);
 EXTERN  INIT(   int     ParseUntriggered, 1);
 
 EXTERN	char	const *InitialFile;
